Modificaciones y mejoras

+++ api/src/controllers/deleteFav.js
@@ -4,11 +4,18 @@ const deleteFav = async (req, res) => {
     try {
         const { id } = req.params;
 
-        await Favorite.destroy({ where: { id } })
+        const favId = parseInt(id);  // Validar que id sea un número entero.
+        if (isNaN(favId)) {
+            return res.status(400).json({ error: 'El ID proporcionado no es válido' })
+        }
+
+        await Favorite.destroy({ where: { id } });  // Eliminar el favorito con el ID proporcionado.
         const allFavorites = await Favorite.findAll()
-        return res.status(200).json(allFavorites)
+
+        return res.status(200).json({ message: 'Favorito eliminado correctamente', allFavorites })
     } catch (error) {
-        return res.status(500).send(error.message)
+        console.error('Error al eliminar favorito:', error);
+        return res.status(500).json({ error: 'Hubo un error al eliminar el favorito' });
     }
 };
